+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/model/Category.java
@@ -1,6 +1,6 @@
 package ru.dpopkov.knowthenics.model;
 
-public class Category {
+public class Category extends BaseEntity {
     private String name;
     private String description;