+++ History.md
@@ -2,6 +2,7 @@
 
 #### 2020-12-11
 * Created Interfaces for KnowThenics Services. Closes #4.
+* Create BaseEntity to add ID to model objects. Closes #5.
 
 #### 2020-12-09
 * Created multi-module project. Closes #1.